backend/pyproject.toml
[tool.poetry]
name = "reduck"
version = "0.5.13"
description = "Data Redaction Application"
authors = ["Ceyhun Kerti <ceyhun.kerti@bluecolor.io>"]
packages = [
    { include="app", from="." }
]
include=["app/public"]
readme="README.md"
repository="https://github.com/bluecolor/duck"
license="CC BY-NC-SA 4.0"
documentation="https://bluecolor.github.io/duck/"
keywords=["redaction", "masking"]


[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
arq = "^0.19"
sqlalchemy-utils = "^0.36.5"
alembic = "^1.4.2"
psycopg2-binary = "^2.8.5"
uvicorn = "^0.11.5"
email_validator = "^1.1.0"
sentry-sdk = "^0.14.3"
click = "^7.1.2"
passlib = "^1.7.4"
cx-Oracle = "^8.1.0"
cryptography = "^3.4.4"
fastapi-pagination = "^0.6.1"
celery = "^5.0.5"
redis = "^3.5.3"
pydash = "^4.9.3"
aiofiles = "^0.6.0"
python-multipart = "^0.0.5"
python-jose = "^3.2.0"
fastapi-mail = "^0.3.5"
aioredis = "^1.3.1"
async-timeout = "^3.0.1"
starlette-context = "^0.3.1"
mkdocs-material = "^7.0.6"
Jinja2 = "^2.11.3"
"keyrings.alt" = "^4.0.2"
pymssql = "^2.2.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
bandit = "^1.6.2"
flake8 = "^3.8.1"
rope = "^0.17.0"
watchgod = "^0.6"
taskipy = "^1.2.1"
pre-commit = "^2.4.0"
isort = {version = "^4.3.21", extras = ["pyproject"]}

[tool.taskipy.tasks]
# Launch Commands for Local Development
app = "uvicorn app.main:app --reload"
worker = "LD_LIBRARY_PATH=/home/ceyhun/projects/lab/duck/lib/instantclient_12_2 arq app.worker.WorkerSettings --watch ./"
# Docker-Compose Tasks
compose-up = "docker-compose -f docker-compose.local.yml -f docker-compose.worker.yml -f docker-compose.yml up --build"
compose-down = "docker-compose -f docker-compose.local.yml -f docker-compose.worker.yml -f docker-compose.yml down --remove-orphans"

[tool.black]
line-length = 79
target-version = ['py38']
exclude = '''
/(
    \.git
  | .vscode
  | build
  | dist
  | .scripts
  | .git-crypt
  | services/.scripts
  | services/.vscode
  | core/.scripts
  | core/.vscode
)/
'''

[tool.isort]
line_length = 79
multi_line_output = 3
not_skip = "__init__.py"
use_parentheses = true
include_trailing_comma = true
force_grid_wrap = 0
known_third_party = ["alembic", "arq", "fastapi", "pydantic", "sentry_sdk", "sqlalchemy", "sqlalchemy_utils", "starlette"]


[tool.poetry.scripts]
duck = 'app.cli:duck'